# solution/docs/api.yaml
openapi: 3.1.0
info:
  title: Advertisement Service
  description: PROD Advertisement Serviice API
  version: v1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /advertisers/{advertiserId}/campaigns/{advertisementId}:
    get:
      tags:
        - Campaigns
      summary: Получение кампании по ID
      description: Создаёт новую рекламную кампанию для указанного рекламодателя.
      operationId: getById
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Кампания успешно получена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
        "403":
          description: Кампания вам не принадлежит!
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
        "404":
          description: Кампания не найдена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
    put:
      tags:
        - Campaigns
      summary: Обновление рекламной кампании
      description: Обновляет разрешённые параметры рекламной кампании до её старта.
      operationId: updateById
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementEditRequest"
        required: true
      responses:
        "200":
          description: Рекламная кампания успешно обновлена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
        "403":
          description: Кампания вам не принадлежит!
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
        "404":
          description: Кампания не найдена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
    delete:
      tags:
        - Campaigns
      summary: Удаление рекламной кампании
      description: Удаляет рекламную кампанию рекламодателя по заданному campaignId.
      operationId: deleteById
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Рекламная кампания успешно удалена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
        "403":
          description: Кампания вам не принадлежит!
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
        "404":
          description: Кампания не найдена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertisementView"
...
  /advertisers/{advertiserId}/campaigns:
    get:
      tags:
        - Campaigns
      summary: Получение рекламных кампаний рекламодателя c пагинацией
      description: Возвращает список рекламных кампаний для указанного рекламодателя
        с пагинацией.
      operationId: getAllByAdvertiserId
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        "200":
          description: Список рекламных кампаний рекламодателя.
          content:
            '*/*':
              schema:
                type: object
        "204":
          description: Список рекламных кампаний рекламодателя пуст.
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Рекламодатель не найден.
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
        - Campaigns
      summary: Создание рекламной кампании
      description: Создаёт новую рекламную кампанию для указанного рекламодателя.
      operationId: create_1
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementCreateRequest"
        required: true
      responses:
        "201":
          description: Рекламная кампания успешно создана.
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
  /advertisers/{advertiserId}/campaigns/generate:
    post:
      tags:
        - Campaigns
      summary: Сгенерировать описание
      description: Генерирует описание по предоставленному заголовку.
      operationId: generateDescription
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Текст сгенерирован
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultResponseString"
        "503":
          description: Сервис недоступен :(
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultResponseString"
  /advertisers/bulk:
    post:
      tags:
        - Advertisers
      summary: Массовое создание/обновление рекламодателей
      description: Создаёт новых или обновляет существующих рекламодателей
      operationId: createOrUpdateAdvertisers
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AdvertiserRequest"
        required: true
      responses:
        "201":
          description: Успешное создание/обновление рекламодателей
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdvertiserRequest"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdvertiserRequest"
  /ads/{advertisementId}/click:
    post:
      tags:
        - Ads
      summary: Фиксация перехода по рекламному объявлению
      description: Фиксирует клик (переход) клиента по рекламному объявлению.
      operationId: clickAd
      parameters:
        - name: advertisementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientIdRequest"
        required: true
      responses:
        "204":
          description: Переход по рекламному объявлению успешно зафиксирован.
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Реклама ещё не просмотрена!
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Реклама не найдена
          content:
            '*/*':
              schema:
                type: object
...
  /advertisers/{advertiserId}:
    get:
      tags:
        - Advertisers
      summary: Получение рекламодателя по ID
      description: Возвращает информацию о рекламодателе по его ID.
      operationId: getAdvertiserById
      parameters:
        - name: advertiserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Информация о рекламодателе успешно получена.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertiserRequest"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertiserRequest"
        "404":
          description: Рекламодатель не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvertiserRequest"
...
  /ads:
    get:
      tags:
        - Ads
      summary: Получение рекламного объявления для клиента
      description: "Возвращает рекламное объявление, подходящее для показа клиенту\
        \ с учетом таргетинга и ML скора."
      operationId: getAd
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Рекламное объявление успешно возвращено.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdView"
        "400":
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdView"
        "404":
          description: Подходящая реклама не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdView"
components:
  schemas:
    AdvertisementEditRequest:
      type: object
      properties:
        impressions_limit:
          type: integer
          format: int32
          minimum: 1
        clicks_limit:
          type: integer
          format: int32
          minimum: 0
        cost_per_impression:
          type: number
          format: double
          minimum: 0
        cost_per_click:
          type: number
          format: double
          minimum: 0
        ad_title:
          type: string
        ad_text:
          type: string
        start_date:
          type: integer
          format: int32
          minimum: 0
        end_date:
          type: integer
          format: int32
          minimum: 0
        targeting:
          $ref: "#/components/schemas/Targeting"
      required:
        - ad_text
        - ad_title
        - clicks_limit
        - cost_per_click
        - cost_per_impression
        - end_date
        - impressions_limit
        - start_date
    Targeting:
      type: object
      properties:
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - ALL
        age_from:
          type: integer
          format: int32
          minimum: 0
        age_to:
          type: integer
          format: int32
          minimum: 0
        location:
          type: string
    AdvertisementView:
      type: object
      properties:
        campaign_id:
          type: string
          format: uuid
        advertiser_id:
          type: string
          format: uuid
        impressions_limit:
          type: integer
          format: int32
        clicks_limit:
          type: integer
          format: int32
        cost_per_impression:
          type: number
          format: double
        cost_per_click:
          type: number
          format: double
        ad_title:
          type: string
        ad_text:
          type: string
        start_date:
          type: integer
          format: int32
        end_date:
          type: integer
          format: int32
        targeting:
          $ref: "#/components/schemas/Targeting"
        image_url:
          type: string
      required:
        - ad_text
        - ad_title
        - advertiser_id
        - campaign_id
        - clicks_limit
        - cost_per_click
        - cost_per_impression
        - end_date
        - impressions_limit
        - start_date
        - targeting
...
    AdvertisementCreateRequest:
      type: object
      properties:
        impressions_limit:
          type: integer
          format: int32
          minimum: 1
        clicks_limit:
          type: integer
          format: int32
          minimum: 0
        cost_per_impression:
          type: number
          format: double
          minimum: 0
        cost_per_click:
          type: number
          format: double
          minimum: 0
        ad_title:
          type: string
        ad_text:
          type: string
        start_date:
          type: integer
          format: int32
          minimum: 0
        end_date:
          type: integer
          format: int32
          minimum: 0
        targeting:
          $ref: "#/components/schemas/Targeting"
        image_url:
          type: string
      required:
        - ad_text
        - ad_title
        - clicks_limit
        - cost_per_click
        - cost_per_impression
        - end_date
        - impressions_limit
        - start_date
        - targeting
    ResultResponseString:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    AdvertiserRequest:
      type: object
      properties:
        advertiser_id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - advertiser_id
        - name
    ClientIdRequest:
      type: object
      properties:
        client_id:
          type: string
          format: uuid
      required:
        - client_id
...
    AdView:
      type: object
      properties:
        ad_id:
          type: string
          format: uuid
        advertiser_id:
          type: string
          format: uuid
        ad_title:
          type: string
        ad_text:
          type: string
        image_url:
          type: string
      required:
        - ad_id
        - ad_text
        - ad_title
        - advertiser_id